Re-commit of some previous altered files to fix earlier bad commit.
A commit incorrectly introduced a very large file to main, causing some issues with push changes, even after reverting and deleting the file. This commit goes back to the last healthy commit and includes the good changes made since, while excluding the large file. Below are some of the details of the changes.

- Updated the .gitignore
- Updated the README with all information on the project and performance
- Added sample images for the README
- Fixed bug in train_test_split() and added cost visualization to run.py.
- Cleaned up documentation and formatting in several files.